* master:
  typo in import
  adjusted ays client to not duplicate  run entries in db
  Fix for negative int support in Capnp. Needed for some schemas
  Add initial jwt support (#572)
  Don't crash if there are no repos just return empty dict

+++ lib/JumpScale/data/hrd/HRDSchema.py
@@ -285,8 +285,8 @@ class HRDSchema:
         typesmap = {
             'str': 'Text',
             'multiline': 'Text',
-            'int': 'UInt32',
-            'integer': 'UInt32',
+            'int': 'Int64',
+            'integer': 'Int64',
             'bool': 'Bool',
             'float': 'Float32',
         }
